create uistate for bookmarksheet
