+++ feature/sessions/src/main/java/io/github/droidkaigi/confsched2023/sessions/section/BookmarkSheet.kt
@@ -23,17 +23,43 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import io.github.droidkaigi.confsched2023.model.DroidKaigi2023Day
 import io.github.droidkaigi.confsched2023.model.TimetableItem
-import io.github.droidkaigi.confsched2023.sessions.BookmarkScreenUiState
-import io.github.droidkaigi.confsched2023.sessions.BookmarkScreenUiState.Empty
-import io.github.droidkaigi.confsched2023.sessions.BookmarkScreenUiState.ListBookmark
+import io.github.droidkaigi.confsched2023.model.TimetableItemId
 import io.github.droidkaigi.confsched2023.sessions.SessionsStrings.BookmarkedItemNotFound
 import io.github.droidkaigi.confsched2023.sessions.SessionsStrings.BookmarkedItemNotFoundSideNote
 import io.github.droidkaigi.confsched2023.sessions.component.BookmarkFilters
+import io.github.droidkaigi.confsched2023.sessions.section.BookmarkSheetUiState.Empty
+import io.github.droidkaigi.confsched2023.sessions.section.BookmarkSheetUiState.ListBookmark
+import kotlinx.collections.immutable.PersistentList
+import kotlinx.collections.immutable.PersistentMap
+import kotlinx.collections.immutable.PersistentSet
+
+sealed interface BookmarkSheetUiState {
+    val currentDayFilter: PersistentList<DroidKaigi2023Day>
+    val isAll: Boolean
+        get() = currentDayFilter.size == DroidKaigi2023Day.values().size
+    val isDayFirst: Boolean
+        get() = currentDayFilter.size == 1 && currentDayFilter.first() == DroidKaigi2023Day.Day1
+    val isDaySecond: Boolean
+        get() = currentDayFilter.size == 1 && currentDayFilter.first() == DroidKaigi2023Day.Day2
+    val isDayThird: Boolean
+        get() = currentDayFilter.size == 1 && currentDayFilter.first() == DroidKaigi2023Day.Day3
+
+    data class Empty(
+        override val currentDayFilter: PersistentList<DroidKaigi2023Day>,
+    ) : BookmarkSheetUiState
+
+    data class ListBookmark(
+        val bookmarkedTimetableItemIds: PersistentSet<TimetableItemId>,
+        val timetableItemMap: PersistentMap<String, List<TimetableItem>>,
+        override val currentDayFilter: PersistentList<DroidKaigi2023Day>,
+    ) : BookmarkSheetUiState
+}
 
 @Composable
 fun BookmarkSheet(
-    uiState: BookmarkScreenUiState,
+    uiState: BookmarkSheetUiState,
     scrollState: LazyListState,
     onTimetableItemClick: (TimetableItem) -> Unit,
     onBookmarkClick: (TimetableItem) -> Unit,
